refactor(controller): change how controller endpoints detect logged user

+++ src/main/java/showtime_corp/profile_vitaile/repository/RoadMapRepository.java
@@ -2,12 +2,11 @@ package showtime_corp.profile_vitaile.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import showtime_corp.profile_vitaile.entity.Employability;
 import showtime_corp.profile_vitaile.entity.RoadMap;
 
 import java.util.Optional;
 
 @Repository
-public interface RoadMapRepository extends JpaRepository<RoadMap, Integer> {
-    Optional<RoadMap> findByUserId(Integer userId);
+public interface RoadMapRepository extends JpaRepository<RoadMap, Long> {
+    Optional<RoadMap> findByUserId(Long userId);
 }
